Fix SSR event listener registration
