+++ packages/ui/.changes/patch.skip-ssr-event-listeners.md
@@ -1,0 +1,1 @@
+Make `addEventListeners()` a no-op during server rendering so SSR signals are not passed to brand-checking runtimes such as Cloudflare Workers and listeners are not retained across requests.
